@@ -52,45 +16,68 @@
     "type": "git",
     "url": "https://github.com/thanarie/html-extract-data.git"
   },
+  "main": "./index.js",
+  "types": "./index.d.ts",
+  "files": [
+    "lib/**/*",
+    "index.js",
+    "index.d.ts",
+    "README.md"
+  ],
+  "scripts": {
+    "prepublishOnly": "npm-run-all -s validate build",
+    "validate": "npm-run-all -p lint test",
+    "dev": "npm-run-all -p dev:*",
+    "dev:ts": "tsc --noEmit --allowJs --watch",
+    "build": "npm-run-all -s clean build:* clean:build",
+    "build:ts": "tsc -p ./tsconfig.build.json && shx cp -Rf decl/* . && shx cp -Rf tmp/* .",
+    "clean": "npm-run-all clean:*",
+    "clean:test": "shx rm -rf coverage .nyc_output",
+    "clean:build": "shx rm -rf tmp decl",
+    "clean:npm": "shx rm -rf lib index.js index.d.ts",
+    "test": "jest ./src/",
+    "test:dev": "jest ./src/ -w",
+    "lint": "npm-run-all lint:*",
+    "lint:js": "eslint src --cache",
+    "prettify": "prettier --write \"src/**/*.{js,ts,json}\"",
+    "precommit": "lint-staged"
+  },
+  "lint-staged": {
+    "gitDir": "./",
+    "linters": {
+      "src/**/*.{js,ts,json}": [
+        "prettier --write",
+        "git add"
+      ],
+      "src/**/*.ts": [
+        "npm run lint:js"
+      ]
+    }
+  },
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.35",
-    "@babel/core": "^7.0.0-beta.35",
-    "@babel/preset-env": "^7.0.0-beta.35",
-    "@babel/preset-stage-3": "^7.0.0-beta.35",
-    "@babel/preset-typescript": "^7.0.0-beta.35",
-    "@babel/register": "^7.0.0-beta.35",
-    "@types/mocha": "^2.2.44",
-    "babel-eslint": "^8.0.3",
-    "babel-plugin-istanbul": "^4.1.5",
-    "chai": "^4.1.2",
-    "cross-env": "^5.1.1",
-    "eslint": "^4.13.1",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-prettier": "^2.4.0",
-    "husky": "^0.14.3",
-    "jsdom": "^11.5.1",
-    "jsdom-global": "^3.0.2",
-    "lint-staged": "^6.0.0",
-    "mocha": "^4.0.1",
-    "npm-run-all": "^4.1.2",
-    "nyc": "^11.3.0",
-    "prettier": "^1.9.2",
-    "shx": "^0.2.2",
-    "ts-node": "^4.0.2",
-    "tslint": "^5.8.0",
-    "tslint-config-airbnb": "^5.4.2",
-    "tslint-config-prettier": "^1.6.0",
-    "typescript": "^2.6.2"
+    "@types/jest": "^26.0.15",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.5.0",
+    "eslint-config-airbnb": ">= 18",
+    "eslint-config-airbnb-base": ">= 13",
+    "eslint-config-prettier": ">= 4",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-import-resolver-webpack": "^0.12.2",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.3.0",
+    "jest": "^26.3.0",
+    "lint-staged": "^10.5.2",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.2.0",
+    "shx": "^0.3.3",
+    "ts-jest": "^26.2.0",
+    "typescript": "^4.1.2"
   },
   "dependencies": {
-    "@types/joi": "^13.0.3",
-    "@types/lodash": "^4.14.91",
-    "detect-node": "^2.0.3",
-    "joi": "^13.0.2",
-    "joi-browser": "^13.0.1",
-    "lodash": "^4.17.4"
+    "@types/lodash-es": "^4.17.3",
+    "lodash-es": "^4.17.15"
   }
 }
