@@ -1,7 +1,5 @@
-import { extractData } from '../../src/lib/extract-data';
-import { createHTML } from '../helpers';
-
-const expect = require('chai').expect;
+import { extractData } from './extract-data';
+import { createHTML } from './testUtils';
 
 describe('extractDataFromElement', () => {
   it('should extract text', () => {
@@ -10,7 +8,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'title';
     const actual = extractData(html, 'h2');
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
   it('should extract text as object config', () => {
     const html = createHTML(`<div>
@@ -18,7 +16,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'title';
     const actual = extractData(html, { query: 'h2' });
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract text when html', () => {
@@ -27,7 +25,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'title';
     const actual = extractData(html, 'h2');
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
   it('should extract text when html as object config', () => {
     const html = createHTML(`<div>
@@ -35,7 +33,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'title';
     const actual = extractData(html, { query: 'h2' });
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract html', () => {
@@ -44,7 +42,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'description <b>bold</b>';
     const actual = extractData(html, 'p', { html: true });
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract html as object config', () => {
@@ -52,8 +50,8 @@ describe('extractDataFromElement', () => {
       <p>description <b>bold</b></p>
     </div>`);
     const expected = 'description <b>bold</b>';
-    const actual = extractData(html, { query: 'p',  html: true });
-    expect(actual).to.equal(expected);
+    const actual = extractData(html, { query: 'p', html: true });
+    expect(actual).toEqual(expected);
   });
 
   it('should extract attribute', () => {
@@ -62,7 +60,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = 'foo.jpg';
     const actual = extractData(html, 'img', { attr: 'src' });
-    expect(actual).to.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract a list', () => {
@@ -75,7 +73,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = ['a', 'b', 'c'];
     const actual = extractData(html, '.tags > .tag', { list: true });
-    expect(actual).to.deep.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract a list and fetch attributes', () => {
@@ -88,7 +86,7 @@ describe('extractDataFromElement', () => {
     </div>`);
     const expected = ['1', '2', '3'];
     const actual = extractData(html, '.tags > .tag', { list: true, attr: 'data-id' });
-    expect(actual).to.deep.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract a list and fetch attributes and text as object', () => {
@@ -108,7 +106,7 @@ describe('extractDataFromElement', () => {
       list: true,
       data: { id: 'data-id', text: true },
     });
-    expect(actual).to.deep.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should extract a list and convert the fetched attributes to numbers', () => {
@@ -120,12 +118,12 @@ describe('extractDataFromElement', () => {
       </ul>
     </div>`);
     const expected = [1, 2, 3];
-    const actual = extractData(
-      html,
-      '.tags > .tag',
-      { list: true, attr: 'data-id', convert: 'number' }
-    );
-    expect(actual).to.deep.equal(expected);
+    const actual = extractData(html, '.tags > .tag', {
+      list: true,
+      attr: 'data-id',
+      convert: 'number',
+    });
+    expect(actual).toEqual(expected);
   });
 
   it('should extract a list and fetch attributes and text as object, and convert to number', () => {
@@ -145,25 +143,27 @@ describe('extractDataFromElement', () => {
       list: true,
       data: {
         id: { attr: 'data-id', convert: 'number' },
-        text: true
+        text: true,
       },
     });
-    expect(actual).to.deep.equal(expected);
+    expect(actual).toEqual(expected);
   });
 
   it('should throw in missing first parameter', () => {
     const actual = () => {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore: intentional
       extractData();
     };
-    expect(actual).to.throw('Missing first parameter \'parent\'');
+    expect(actual).toThrow("Missing first parameter 'parent'");
   });
 
   it('should throw in missing second parameter', () => {
     const actual = () => {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore: intentional
       extractData(document.createElement('div'));
     };
-    expect(actual).to.throw('Missing second parameter \'config\'');
+    expect(actual).toThrow("Missing second parameter 'config'");
   });
 });
