@@ -1,36 +1,79 @@
-// http://eslint.org/docs/user-guide/configuring
-
-module.exports = {
-  root: true,
-  parser: 'babel-eslint',
+var typeScriptSettings = {
+  parser: '@typescript-eslint/parser',
   parserOptions: {
-    sourceType: 'module',
+    project: './tsconfig.json',
   },
-  env: {
-    browser: true,
-  },
-  extends: ['airbnb-base', 'prettier'],
-  plugins: ['import', 'prettier'],
+  extends: [
+    'plugin:@typescript-eslint/recommended',
+    'prettier/@typescript-eslint',
+    'plugin:prettier/recommended',
+    'plugin:import/typescript',
+  ],
   rules: {
-    'import/extensions': [
+    '@typescript-eslint/array-type': ['error', { default: 'generic', readonly: 'generic' }],
+    '@typescript-eslint/ban-ts-comment': 'error',
+    '@typescript-eslint/consistent-type-assertions': 'error',
+    '@typescript-eslint/explicit-member-accessibility': 'error',
+    '@typescript-eslint/indent': 'off',
+    '@typescript-eslint/naming-convention': [
       'error',
-      'always',
-      // hide known extensions that are resolved by webpack
       {
-        js: 'never',
-        ts: 'never',
+        selector: 'default',
+        format: ['camelCase'],
+        leadingUnderscore: 'forbid',
+        trailingUnderscore: 'forbid',
+      },
+      {
+        selector: 'typeLike',
+        format: ['PascalCase'],
+      },
+      {
+        selector: 'variable',
+        format: ['camelCase', 'UPPER_CASE', 'PascalCase'],
+      },
+      {
+        selector: 'enumMember',
+        format: ['PascalCase'],
       },
     ],
-    // prettier compatibility
-    'max-len': ['error', 100],
-    'prettier/prettier': [
+    '@typescript-eslint/no-empty-function': 'error',
+    '@typescript-eslint/no-explicit-any': 'error',
+    '@typescript-eslint/no-inferrable-types': 'off',
+    '@typescript-eslint/no-unused-vars': [
       'error',
-      { singleQuote: true, trailingComma: 'all', printWidth: 100, tabWidth: 2 },
+      { vars: 'all', args: 'after-used', ignoreRestSiblings: true, varsIgnorePattern: "Fragment" },
     ],
-    // only for use with getter-setters
-    'no-underscore-dangle': 0,
-    // to correctly work on windows with some tools that create windows line-endings
-    // this will be correct by git when committed
-    'linebreak-style': 0
+    'no-use-before-define': 'off',
+    '@typescript-eslint/no-use-before-define': ['error', { functions: false }],
+    '@typescript-eslint/prefer-interface': 'off',
+    '@typescript-eslint/prefer-readonly': 'error',
+    camelcase: 'off',
   },
+};
+
+module.exports = {
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+  },
+  plugins: ['import'],
+  env: {
+    browser: true,
+    es6: true,
+  },
+  globals: {
+    require: true,
+    process: true,
+  },
+  settings: {
+  },
+  rules: {
+    'no-console': 'warn',
+  },
+  overrides: [
+    {
+      files: ['*.ts'],
+      ...typeScriptSettings,
+    },
+  ],
 };
