@@ -1,9 +1,19 @@
-import mergeWith from 'lodash/mergeWith';
-import { extractData, getSchema } from './extract-data';
-import { extractFromElement, getDataSchema } from './extract-from-element';
-import { validateSchema } from './validation';
+import mergeWith from 'lodash-es/mergeWith';
+import { extractData } from './extract-data';
+import { extractFromElement } from './extract-from-element';
+import type { ExtractConfig, ExtractDataConfig } from './types';
 
-export function extractFromHTML(container: HTMLElement, config: any, additionalData: any = {}) {
+/**
+ * Top level function, does input parameter checking, and splits off between single and list mode
+ * @param container
+ * @param config
+ * @param additionalData
+ */
+export function extractFromHTML(
+  container: HTMLElement,
+  config: ExtractConfig,
+  additionalData: Record<string, unknown> = {},
+): Record<string, unknown> | Array<Record<string, unknown>> {
   if (!container) {
     throw new Error("Missing first parameter 'container'");
   }
@@ -11,30 +21,39 @@ export function extractFromHTML(container: HTMLElement, config: any, additionalD
     throw new Error("Missing second parameter 'config'");
   }
 
-  const validatedConfig = validateSchema(config, schema);
-
-  if (validatedConfig.list) {
-    const elements = Array.from(container.querySelectorAll(validatedConfig.query));
-    return elements.map((element: HTMLElement) =>
-      getData(element, validatedConfig, additionalData),
-    );
+  // the `query` from the config is used to select children and process each one individually
+  if (config.list) {
+    if (!config.query) {
+      throw new Error('When specifying "list", you must also provide a "query".');
+    }
+    const elements = Array.from(container.querySelectorAll<HTMLElement>(config.query));
+    return elements.map((element: HTMLElement) => getData(element, config, additionalData));
   }
 
-  return getData(
-    validatedConfig.query ? container.querySelector(validatedConfig.query) : container,
-    validatedConfig,
-    additionalData,
-  );
+  // if not a list, this will extract all information from either the single element specified
+  // in the query, or the passed root element itself
+  const element = config.query ? container.querySelector<HTMLElement>(config.query) : container;
+  return getData(element, config, additionalData);
 }
 
-function getData(element, config, additionalData) {
-  const extractedData = Object.keys(config.data).reduce((obj, key) => {
-    const dataConfig = config.data[key];
+/**
+ * Goes through the 'self' and 'data' config to extract information for each key.
+ * - for the data, it can either be a config object or a function
+ * - for self, it can only be from this element
+ * @param element
+ * @param config
+ * @param additionalData
+ */
+function getData(
+  element: HTMLElement | null,
+  config: ExtractConfig,
+  additionalData: Record<string, unknown>,
+) {
+  const extractedData = Object.entries(config.data || {}).reduce((obj, [key, dataConfig]) => {
     let extracted;
     if (typeof dataConfig === 'function') {
       extracted = dataConfig(
-        // @ts-ignore: TS doesn't understand passing destructure args yet
-        (...args) => extractData(element, ...args),
+        (...args) => extractData(element, ...(args as [config: string | ExtractDataConfig])),
         element,
       );
     } else {
@@ -42,7 +61,7 @@ function getData(element, config, additionalData) {
     }
     obj[key] = extracted;
     return obj;
-  }, {});
+  }, {} as Record<string, unknown>);
 
   let selfData;
   if (config.self) {
@@ -57,27 +76,3 @@ function getData(element, config, additionalData) {
     }
   });
 }
-
-const schema = (() => {
-  const isProd = process.env.NODE_ENV === 'production';
-  /* istanbul ignore if */
-  if (isProd) return {};
-
-  const Joi = require('./vendor/joi-browser');
-  return Joi.object()
-    .keys({
-      query: Joi.string(),
-      data: Joi.object()
-        .pattern(/.*/, [
-          Joi.string(),
-          Joi.func()
-            .minArity(1)
-            .maxArity(2),
-          getSchema(),
-        ])
-        .required(),
-      list: Joi.bool(),
-      self: getDataSchema(),
-    })
-    .with('list', 'query');
-})();
